Started trying to improve AST
 - Got rid of all original AST node types
 - Re-implemented the 'BurstASTNode' type
 - Re-implemented the 'BurstLiteralExpressionNode' type

Originally, when I started this, I was just going to remove the 'BurstValueExpressionNode', however, in my doing so, I found that the AST is designed quite poorly.
I still have a lot of work left to do on this; I'm not even sure if the contents of this commit will remain, but it's a start.

Signed-off-by: Gigabyte-Giant <brynden.bielefeld@hotmail.com>